+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # archivald
 
+## 0.1.0
+
+### Minor Changes
+
+- - Added some console colors
+  - Fixed minor bugs
+
+## 0.0.3
+
+### Patch Changes
+
+- fixed async/await issue
+
 ## 0.0.2
 
 ### Patch Changes